requirements.txt
alembic==1.11.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
asyncio-redis==0.16.0
asyncpg==0.28.0
base==0.0.0
certifi==2023.5.7
click==8.1.4
dnspython==2.4.0
ecdsa==0.18.0
email-validator==2.0.0.post2
envparse==0.2.0
exceptiongroup==1.1.2
fastapi==0.100.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
passlib==1.7.4
pluggy==1.2.0
psycopg2==2.9.6
pyasn1==0.5.0
pydantic==2.0.2
pydantic-settings==2.0.1
pydantic_core==2.1.2
pytest==7.4.0
pytest-asyncio==0.21.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
redis==4.6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.18
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.7.1
uvicorn==0.22.0
